Substantial reorganization and refactoring.
Broke functionality into separate files. The model specification is just one file now. Not sure about that, since the parameters still have to be in synch with the rest of the code. I guess that's the danger of specifying a model in python instead of a prototxt. A TODO if this wasn't a toy project.

viewer.py isn't hooked up to the CNN model at this commit.